+++ src/main/resources/templates/layout/sidebar.html
@@ -23,7 +23,8 @@
                     <div><a href="#" class="text-white text-sm" th:onclick="regitClient()">- 고객정보 입력</a></div>
                     <div><hr></div>
                     <div><a href="/user/employee/emmain" class="text-white text-sm">- 직원정보 보기</a></div>
-                    <div><a href="/user/employee/emform" class="text-white text-sm" th:onclick="regitEmployee()">- 직원정보 입력</a></div>
+<!--                    20230207 15:44 수정-->
+                    <div><a href="#" class="text-white text-sm" th:onclick="regitEmployee()">- 직원정보 입력</a></div>
                     <div><hr></div>
                     <div><a href="#" class="text-white text-sm">- 장비정보 보기</a></div>
                     <div><a href="#" class="text-white text-sm" th:onclick="regitMachine()">- 장비정보 입력</a></div>
