+++ src/main/resources/templates/user/employee/emeditform.html
@@ -37,7 +37,7 @@
 <div class="card m-3">
 
   <div class="card-body">
-    <h5 class="card-title mt-3 mb-3">인사 정보 등록</h5>
+    <h5 class="card-title mt-3 mb-3">인사 정보 상세보기</h5>
     <div class="row">
       <div class="col-sm-12">
         <div class="form-group">
@@ -136,15 +136,18 @@
         </div>
       </div>
 
-      <input type="button" class="btn btn-primary" id="empSubmitBtn" value="등록하기" />
+
+
+      <input type="button" class="btn btn-primary" id="empSubmitBtn" value="등록하기"/>
     </div>
   </div>
 </div>
-
 <script>
-  const empSubmitBtn = document.querySelector('#empSubmitBtn');
 
-    empSubmitBtn.addEventListener("click", () => {
+  const empSubmitBtn = document.querySelector('#empSubmitBtn');
+  const urlParams = new URL(location.href).searchParams;
+
+  empSubmitBtn.addEventListener("click", () => {
     // const input_emp_id = document.querySelector('#emp_id').value;
     const input_cp_name = document.querySelector('#cp_name').value;
     const input_emp_name = document.querySelector('#emp_name').value;
@@ -175,28 +178,29 @@
       emp_etc3 : input_emp_etc3
     };
 
-    if(input_cp_name == "" ||
-            input_emp_name == "" || input_emp_joining == "" ||
-            input_emp_regid == ""){
+    if (input_misu_div == "" || input_cp_name == "" ||
+            input_cst_code == "" || input_misu_manager == "" ||
+            input_misu_value == "" || input_misu_name == "" ||
+            input_misu_total_value == "") {
       alert('*이 있는 칸은 채워주시기 바랍니다.')
       return;
     }
 
-    console.log(emp1);
+    console.log(misu1);
 
-    const url = "/user/employee/emform";
+    const url = "/api/admin/misu";
 
     fetch(url, {
-      method : "POST",
-      body : JSON.stringify(emp1),
-      headers : {
-        "Content-Type" : "application/json"
+      method: "PUT",
+      body: JSON.stringify(misu1),
+      headers: {
+        "Content-Type": "application/json"
       }
     }).then(response => {
-      if(response.ok){
+      if (response.ok) {
         response.text();
         alert('등록이 완료되었습니다.');
-        window.location.reload();
+        window.close();
       } else {
         alert('등록을 실패하였습니다.');
       }
